# src/main/java/br/org/portalfadesp/servicopagamento/domain/pagamento/exception/ErrorResponse.java
package br.org.portalfadesp.servicopagamento.domain.pagamento.exception;

import com.fasterxml.jackson.annotation.JsonFormat;
import java.time.LocalDateTime;

public class ErrorResponse {

  private String mensagem;
  private int statusCode;

  @JsonFormat(pattern = "dd/MM/yyyy HH:mm:ss")
  private LocalDateTime timestamp;

  private String path;

  public ErrorResponse(String mensagem, int statusCode, String path) {
    this.mensagem = mensagem;
    this.statusCode = statusCode;
    this.timestamp = LocalDateTime.now();
    this.path = path;
  }

  public String getMensagem() {
    return mensagem;
  }

  public void setMensagem(String mensagem) {
    this.mensagem = mensagem;
  }

  public int getStatusCode() {
    return statusCode;
  }

  public void setStatusCode(int statusCode) {
    this.statusCode = statusCode;
  }

  public LocalDateTime getTimestamp() {
    return timestamp;
  }

  public void setTimestamp(LocalDateTime timestamp) {
    this.timestamp = timestamp;
  }

  public String getPath() {
    return path;
  }

  public void setPath(String path) {
    this.path = path;
  }
}